refactor: move CheckoutForm and CheckoutPage
to Checkout/components folder

Other Changes:
- Updated imports
- Move test files to Checkout/components/test
