+++ src/domains/billing/Checkout/hooks/useTaskFetch.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import fetch from "node-fetch";
-import { createApiEvent } from "../CheckoutForm";
 import { useApiEvents } from "src/common/ApiResponse/ApiEventsContext";
+import { createApiEvent } from "../utils";
 
 export const getTaskURL = `${process.env.NEXT_PUBLIC_API_URL}api/v1/tasks`;
