+++ src/domains/billing/Checkout/utils/index.tsx
@@ -1,0 +1,20 @@
+import { v4 as uuid } from "uuid";
+
+import { ApiEvent } from "src/common/ApiResponse";
+
+export function createApiEvent({
+  id = uuid(),
+  message,
+}: {
+  id?: string;
+  message: string;
+}): ApiEvent {
+  return {
+    id: id,
+    isError: true,
+    title: message,
+    status: 500,
+    statusText: "",
+    level: "error",
+  } as const;
+}
